refactor: rename JsonWebKey types
