+++ packages/shared/src/index.ts
@@ -11,4 +11,4 @@ export * from './TypedMessage'
 export * from './utils'
 export * from './hooks'
 export * from './database'
-export * from './modules'
+export * from './crypto/JWKType'
