--- packages/shared/src/modules/index.ts
@@ -1,1 +1,0 @@
-export * from './CryptoAlgorithm/interface/utils'
